fix(core): check position in Merkle tree when verifying proofs (#349)
So far, our Merkle proofs only checked if some leaf was included in a
Merkle tree, but not its *position* in the tree. In all our use cases,
we need that position check as well.

This commit changes the verification such that it takes the leaf index as an
input.