refactor: remove `std` from `walrus-core` (#364)
Most code changes are simply changing an `std::` import to a `core::` or
`alloc::` import.

We still use `std` for `std::error::Error`, so this doesn't make
`walrus-core` a true `no_std` crate. However, `std::error::Error` is
supported on WASM, which is the reason for this change.

Also add lints to encourage importing resources from `alloc` and `core`
instead of `std` (if available).