fix: force git checkout (#436)
Basically Cargo.lock has been modified on the remote machines and so git
checkout origin/main failed. This can happen if we compile/run the
binary manually on the remote machine (without using the orchestrator)
or perhaps if we brutally interrupted the orchestrator while it was
updating / recompiling the code (not sure though). This problem should
now be fix by forcing git to checkout origin/main