fix: Make delete blob an atomic op (#467)
This PR makes the two delete operations i.e. `delete_metadata` and
`delete_slivers` atomic. While i don't think this is a big issue as
these operations would be idempotent if there was a server restart
between the first and second delete because the event is going to get
reprocessed but it could still lead to some potential confusing state
where the slivers exist but its metadata doesn't (or even more confusing
would be entry in blob_info cf exist but not in metadata cf)