feat(client): add blocklist feature for daemon (#539)
The blocklist is simply a list of blob IDs in YAML syntax and can be set
with an optional argument. This is currently only read once during
startup. Continuous updates could be added in a follow-up PR.

Both aggregator and publisher return a 403 (Forbidden) message for
blocked IDs.

The functionality is added to the `Client`, so it would also work for
the CLI `store` and `read`. However, I don't really see a need for this
and thus haven't implemented it.