feat: Add crash recovery testing (#530)
The script works in configurable number of runs. In each run a cluster
is brought up (by copying data dir of previous run), some storage nodes
are crashed and a file is stored. In the end, we attempt to read all
blobs stored and compare.
Run it as:
```
 ./scripts/crash_recovery.sh  
```
from the walrus directory