feat: resolve `~` to home directory in paths (#563)
So far, this only worked in direct CLI options. With this change, it
also works for the JSON API and for paths in config files.

For example, one can now set a wallet config path in the
`client_config.yaml` as follows:

```yaml
wallet_config: ~/.sui/sui_config/client.yaml
```

Also, the JSON example now works, which didn't before:

```sh
walrus json '{
    "config": "~/.walrus/client_config.yaml",
    "command": {"blobId": {"file": "~/.walrus/client_config.yaml"}}
}'
```

The `~` is only replaced if it occurs as a directory at the beginning of
paths.