feat(aggregator): add cache-friendly headers (#456)
Added headers:

- `Cache-control: public, max-age=86400, stale-while-revalidate=3600`
- `ETag: <BLOB_ID>`

Rationale:

Originally I had considered adding `Cache-control: immutable,...`.
This header ensures that the resource is revalidated (queried from the
server) only if it is stale (`max-age` is expired). However, this is
not great for blob invalidation.

Therefore, I opted for not having `immutable`, so the browser checks
if the resource has changed by sending the get request with
the `ETag`. However, this requires the aggregator to respond 
`304 not modified` in case the blob has not been invalidated.

Therefore, this PR currently depends on #422 . Once that is done,
the aggregator can check if there is an `Etag` header in the request,
query the status of the blob, and respond `304 not modified` if it
has not been invalidated.

`stale-while-revalidate` ensures that expired resources can be
revalidated asynchronously.

Closes #454

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>