refactor!: reorganize `walrus-service` crate (#625)
This reorganizes the `walrus-service` modules and adds new Cargo
features.

We now have five top-level modules activated with different features:
- `common` contains functionality shared by client and storage node.
Essentially some API definitions, telemetry settings, and other
utilities. This is activated with either of the `client` and `node`
features.
- `client` contains all client-related code (including the client
daemon) and is activated with the `client` feature.
- `node` contains all storage-node-related code and is activated with
the `node` feature.
- `testbed` contains all functionality for creating configuration and
deploying the system contract and is activated with the `deploy`
feature.
- `test_utils` contains various test utilities and is activated with the
`test-utils` feature.

In addition, this moves the admin functionality from the `walrus-node`
CLI to a new `walrus-deploy` CLI and moves the end-to-end tests to a new
`walrus-e2e-tests` crate. The latter is mainly to make sure all
necessary features are activated for the tests.

Finally, this contains a few other changes:
- Change visibility of some structs and functions, required by the new
crate structure. This required also some new docstrings.
- Some changes to the `api_success_alias` and `rest_api_error` macros.
- Move some definitions into `walrus_service::utils`.
- Clean up some dependencies.
- Clean up some imports, linter exceptions, and docstrings.
- Keep all dependency versions except for the orchestrator in the
workspace's `Cargo.toml` file.
- Create a new example for the node configuration file.
- Clean up some TODOs.

This PR represents a breaking change in the sense that the `walrus-node`
binary no longer provides the administrative functionality. Scripts and
CI workflows that use that functionality need to migrate to
`walrus-deploy`.