feat: only handle sync shard request if src node also in the new epoch (#708)
Previously, we don't check if the source and destination storage nodes
are in the same epoch or not. The idea was that we want the source node
to start sending slivers to the destination even if it hasn't received
epoch change event. This means that the source node may not see all
certified blob up until the epoch boundary yet.

This optimization adds quite a bit complexity, so in this PR, we enforce
that the source storage node should also see the epoch change event, and
has moved to the new epoch before serving any sync shard request. For
honest node, this means that it contains the full list of certified
blobs and the bulk shard transfer should just send all certified blobs'
shard to the destination node. This also makes bulk shard transfer easy
to reason about.

For the optimization mentioned earlier, we can implement it later if
there is a larger delay if epoch change events are received by different
storage node.

Contribute to #259