feat: Integrate event processor with existing walrus code (#745)
The PR integrates new checkpoint based event processor with existing
walrus code. I am working on adding tests but PR is ready for review
(and it has gotten big already), so wanted to get some signal early on.
The biggest underlying change is to replace `usize` with `u128` for
tracking event cursor index in the current codebase. This is required
because with the new event processor, we assign an event sequence number
(u128) to each event which is comprised of two u64 parts:
`CheckpointSequenceNumber` (which is the sequence number of the
checkpoint the event is part of) and `counter` (index of the event in
the sequence of emitted events in a checkpoint). This event sequence
number will be further used to track the event index up to which event
blobs have been certified by the committee in the new endpoint
`certify_event_blob` in the walrus smart contract which we are yet to
add