feat(client): track committee changes (#866)
The client now uses the correct committee when reading and storing
blobs.

Main changes:

- Refactored the `NodeCommunicationFactory` into its own module; it now
takes a `certified_epoch` parameter when creating the vector of
`NodeCommunication` objects to read, so that the correct committee for
reading is selected.
- The Walrus `Client` now owns a (Sui) `ReadClient` for reads. This is
necessary as, before any read, we need to know when the blob was
certified. This is taken from the blob status, which may be read from
chain.
- This latter change also required making `ReadClient` a supertrait of
`ContractClient`. Otherwise, the "store" client would not be able to
read as well.
- Added retries to `get_verified_blob_status`.

Contributes to #753

---------

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>
Co-authored-by: Markus Legner <mlegner@users.noreply.github.com>