fix: adds a limit by the number of slivers being recovered (#916)
This allows limiting by the number of slivers being recovered. It uses
an iterator to create the sliver recovery futures, thereby avoiding
reserving space for all of the futures in advance. As permits are
available, it dispatches the future and awaits the result.

This does not resolve the issue of the amount of memory required for
syncing, however, it gives us another knob to tune for seeing the
performance of the current storage node deployment and maintaining
system stability; to keep a more constant rate of slivers being
recovered (as opposed to needing to wait for the recovery of all of the
slivers for a given blob before being able to start on another set).