fix: lock the wallet before reading the object refs (#1014)
Locking the wallet before signing the transaction prevents race
conditions and locked objects. However, since the object references (and
importantly the object version) were retrieved before the wallet was
locked. It could then appen that two separate transactions would race to
use the same object, and one of the two would fail because the object
version was bumped by the other in the meantime.

This PR ensures that the wallet is passed explicitly to the
`sign_and_send_ptb` function, and that it is locked before any object
ref is taken.

---------

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>