fix(events): more thorough verification of events in sui checkpoints (#826)
When reading downloaded checkpoint for events, we ensure that it is the
right checkpoint summary (by verifying it is signed by the right
committee and points to the previous verified checkpoint) but do not
check checkpoint's content and events. This PR fixes this issue and adds
the missing event verification.