fix: sending epoch_sync_done should not block event processing (#1155)
This PR implements that sending `epoch_sync_done` does not block next
event processing.

We renamed `BackgroundShardRemover` to `StartEpochChangeFinisher`, which
in addition to removing shards, also sends `epoch_sync_done` message.

Note that it possible that `epoch_sync_done` is not executed
successfully, but the next epoch arrives. In this case,
`epoch_sync_done` would abort due to unexpected epoch, and the event
should be able to complete.