refactor: move events into walrus-service and create walrus-utils crate (#1101)
Today, `EventProcessor` is susceptible to Sui FullNode failures and
would crash the binary on failure to download checkpoints and events
after sometime. But, as we recently saw the full nodes can go down for
short periods causing a Walrus outage. But, we shouldn't crash storage
nodes because:
1. Walrus can still serve reads
2. Storage nodes which are lagging have potentially already downloaded
events to their local db and can safely continue with event recovery
even if newer events are not being fetched

This PR does a couple things:
1. Moves `event_processor.rs` into `walrus_service` crate so we can use
`ExponentialBackoff`. `EventProcessor` needs to access `EventBlobs` for
a future PR, so this cyclic dependency needs to be broken
2. Uses exponential backoff to retry checkpoint failures