test: various performance improvements for E2E tests (#893)
This adds a number of changes that help to speed up running integration
tests. For example, they reduce `cargo test --test test_client --
--ignored` from taking around 235 seconds to just 30 seconds:

#### Using an external sui test cluster

An external test cluster started with `sui start` can be used for tests
through the use of the `SUI_TEST_CONFIG_DIR` environment variable. Each
test will still continue to publish it's own walrus packages, but they
will do so to the external test cluster instead of the in-process test
cluster. The cluster needs to already be running.

This removes the time required to build the test cluster, which, for
some tests, was a large portion of their runtime. While there was a
shared in-process test cluster, running many tests against it seemed to
result in errors and retries.

Unfortunately, this does not yet work with `cargo nextest` (despite it
being the motivating reason for not just running `sui start` from inside
the process), as this would require inter-process sharing of the
cluster's wallet config.

#### Use a single transaction to fund wallets in a test

A large amount of time was spent waiting for a wallet to be funded
before releasing the lock on the shared admin wallet to be able to fund
the next wallet. This change funds multiple wallet simultaneously to
avoid the wait.

#### Register storage nodes in parallel

Registers and submits stake for storage nodes in parallel. Reduces each
test's runtime by several seconds.

#### Cache compiled move modules

Since the move code does not change across tests, it only needs to be
compiled once. This change caches the compiled value in a static
variable. This provides speedups when running many tests in parallel
(i.e., 1 test will not benefit; 2 will not really benefit since the
second test which runs in parallel will block waiting on the package to
be compiled by the first test; after a point though, tests are scheduled
and start that already see the package compiled).

Unfortunately, this does not benefit `cargo nextest` as that would
require caching it to the filesystem and doing synchronisation across
the processes.

---------

Co-authored-by: Markus Legner <markus@mystenlabs.com>