feat: [jh/walrus-proxy labels] (#1268)
* [jh/walrus-proxy labels]

## Description
during testing it was discovered we need a way to supply labels in config that grafana-agent previously provided.
this pr adds that ability within walrus-service and also the ability to override them from walrus-proxy, although in a limited fashion.
if we need more robust label management on the proxy, we can explore that in the future but this pr provides 1-1 functionality from what the agent did.

the labels from walrus-node are secondary. if the proxy has the same label in its config, say network=testnet, the walrus-node will not win and will instead receive the proxy's version of that label.  EG, any conflicts between the proxy and the node will default to what the proxy says.  there is where we could improve "overlays" or "overrides" in the future, should the need arise.  If we do that, we will want to look into a more robust label merge process.

## Test Plan
```
scripts/local-testbed.sh -e
```
https://metrics.sui.io/goto/pBHYf1VNR?orgId=1

and a config block to enable it:

```
metrics_push:
  push_url: http://localhost:8080/publish/metrics
  labels:
    host: dryrun-node-0
    network: notnet
    unicorns: i-love-them
```

* straggler lint

* remove some cruft

* lint

* update Cargo.toml

* add liveness

* i am no longer a teapot

* doc lint