feat(cli): mark global options as 'global' (#1474)
This is just a convenience feature, which allows adding these options to
the `walrus` CLI at lower levels. For example, before,
`walrus info --json` caused an error; with the changes it is valid and
behaves the same as `walrus --json info`. Same for
`walrus read <BLOB_ID> --config <CONFIG_PATH>`, etc.