feat: add utility for tracking HTTP request/response bodies (#1731)
This adds the struct `VisitBody` which implements `http_body::Body` and the trait `VisitBody`, which
together can be used to observe polling of an http request/response bodies. This is necessary as
some metrics are measured when the body is consumed.

Additionally, adds several labels to the metrics exported by http servers using the `VisitBody`
type to observe request and response body sizes of each request.