feat(client): allow setting Sui RPC URLs in client config (#2066)
Add a new `rpc_urls` field to the client config, which allows setting
multiple Sui RPC URLs.

When invoking the `walrus` CLI, the Sui RPC URL(s) used for reads are
now selected in the following order:

1. The RPC URL provided as a CLI argument.
2. The RPC URLs set in the client config.
3. The RPC URL set in the wallet config.

If none of these are set, the client will **no longer fall back to the
`fullnode.testnet.sui.io`** but instead exit with an error. This is
because we now have Walrus networks deployed on different Sui networks,
such that this default behavior may cause unexpected behavior.