fix(node): increase limit for number of pending accept reset streams (#2072)
Storage nodes frequently cancel pending requests to other storage nodes
when they are no longer relevant. This triggers a limit within the `h2`
crate, and associated warnings that pollute the logs.

Until we change the behavior of cancelling pending requests, we can
increase the limit to very high values to effectively disable the check.