refactor: bump Rust version to 1.87 and reduce size of some types (#2135)
Clippy emitted a few new warnings `result_large_err` and
`large_enum_variant` and suggested boxing some of the large
variants/types.

This commit wraps some large (and rarely used) enum variants and error
fields in a `Box` to avoid the memory overhead on the stack:
- `walrus_sdk::error::ClientError::kind`
- `walrus_service::node::committee::node_service::Response::InvalidBlobAttestation`
- `walrus_sui::client::SuiClientError::SuiError`

Two other instances of this lint warning are (temporarily) ignored. The
commit also changes some instances of `.iter().any()` to the more
efficient `contains()`.