fix(node): fix continuing event stream from init state (#2273)
There are some problems with the `continue_event_stream` and the
`EventProcessor::events` functions:

The event_stream returned by the continue_event_stream function always
uses the lower of the provided event cursors, and does not necessarily
match the returned `actual_event_index`. Then, in the `process_events`
function, the returned event stream is zipped with an index stream which
starts at the returned `actual_event_index`, so there can be a mismatch.

Furthermore, the `events` function doesn't work correctly when there
aren't any events at the provided index: It doesn't check whether the
events it polls from the event store actually match the expected index
and updating the index doesn't work properly as a result.

What can happen in a node that has empty `events` and
`event_blob_writer` DBs is that the event stream is started at index 0
(because that is the `event_blob_writer_cursor`), while the index stream
is started at the "repositioned" index. The events function then keeps
polling the events store with a low starting index (0, 1000, 2000, ...)
such that always the same events (the first ones in the incomplete
history) are returned and included in the stream.

This commit fixes the problems as follows:

- Modify the `continue_event_stream` function to return an event
  stream matching the returned event cursor.
- Add a consistency check in the `events` function which aborts when
  there is an index mismatch. (This should not actually be triggered,
  it is just a safeguard against further bugs.)