feat(client): add `permanent` flag to CLI and publisher (#2281)
Currently, blobs are "permanent" as default and can be stored as "deletable" through the `--deletable` CLI option and a `deletable` query parameter on the publisher. We want to change this default behavior in the future, at which point we need a way to specify a blob as "permanent". Adding the `--permanent` flag in the same release where the default is changed is a bit dangerous as all users would have to change their invocations at exactly that point.

Adding the `--permanent` flag now allows users to already start calling `walrus store --permanent [...]` to make sure the behavior is unaffected by the later default change. Additionally, we log a warning about the upcoming default change from permanent to deletable if neither flag is specified.