fix(node): cancel blob syncs for expired blobs (#2302)
Since the async event processing, it can happen that a new blob recovery
is started after the epoch-change event for that blob's expiration epoch
is processed. As a result, that blob sync is not cancelled by the task
cancelling blob syncs for expired blobs. This blob sync will then remain
until the next epoch change.

This commit adds expiration checks to the already existing blob sync
task monitor, such that stale blob syncs are cancelled quickly.

It also makes the blob sync task monitor interval configurable, keeping
the current value of 60 seconds as a default.